Moved data reader file to utils.
